+++ app/src/main/kotlin/me/xizzhu/android/joshua/annotated/AnnotatedVersesInteractor.kt
@@ -50,8 +50,8 @@ class AnnotatedVersesInteractor<V : VerseAnnotation>(private val verseAnnotation
     suspend fun bookShortNames(): ViewData<List<String>> =
             viewData { bibleReadingManager.readBookShortNames(currentTranslation()) }
 
-    suspend fun verse(verseIndex: VerseIndex): ViewData<Verse> =
-            viewData { bibleReadingManager.readVerse(currentTranslation(), verseIndex) }
+    suspend fun verses(verseIndexes: List<VerseIndex>): ViewData<Map<VerseIndex, Verse>> =
+            viewData { bibleReadingManager.readVerses(currentTranslation(), verseIndexes) }
 
     suspend fun saveCurrentVerseIndex(verseIndex: VerseIndex) {
         bibleReadingManager.saveCurrentVerseIndex(verseIndex)