+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/BibleReadingManager.kt
@@ -127,8 +127,8 @@ class BibleReadingManager(private val bibleReadingRepository: BibleReadingReposi
                            bookIndex: Int, chapterIndex: Int): List<Verse> =
             bibleReadingRepository.readVerses(translationShortName, parallelTranslations, bookIndex, chapterIndex)
 
-    suspend fun readVerse(translationShortName: String, verseIndex: VerseIndex): Verse =
-            bibleReadingRepository.readVerse(translationShortName, verseIndex)
+    suspend fun readVerses(translationShortName: String, verseIndexes: List<VerseIndex>): Map<VerseIndex, Verse> =
+            bibleReadingRepository.readVerses(translationShortName, verseIndexes)
 
     suspend fun search(translationShortName: String, query: String): List<Verse> =
             bibleReadingRepository.search(translationShortName, query)