+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/BibleReadingRepository.kt
@@ -123,8 +123,8 @@ class BibleReadingRepository(private val localReadingStorage: LocalReadingStorag
                            bookIndex: Int, chapterIndex: Int): List<Verse> =
             localReadingStorage.readVerses(translationShortName, parallelTranslations, bookIndex, chapterIndex)
 
-    suspend fun readVerse(translationShortName: String, verseIndex: VerseIndex): Verse =
-            localReadingStorage.readVerse(translationShortName, verseIndex)
+    suspend fun readVerses(translationShortName: String, verseIndexes: List<VerseIndex>): Map<VerseIndex, Verse> =
+            localReadingStorage.readVerses(translationShortName, verseIndexes)
 
     suspend fun search(translationShortName: String, query: String): List<Verse> =
             localReadingStorage.search(translationShortName, query)