+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/local/LocalReadingStorage.kt
@@ -41,7 +41,7 @@ interface LocalReadingStorage {
     suspend fun readVerses(translationShortName: String, parallelTranslations: List<String>,
                            bookIndex: Int, chapterIndex: Int): List<Verse>
 
-    suspend fun readVerse(translationShortName: String, verseIndex: VerseIndex): Verse
+    suspend fun readVerses(translationShortName: String, verseIndexes: List<VerseIndex>): Map<VerseIndex, Verse>
 
     suspend fun search(translationShortName: String, query: String): List<Verse>
 }